delivery_service/requirements.txt
asgiref==3.8.1
bip-utils==2.9.3
cbor2==5.6.5
cffi==1.17.1
coincurve==21.0.0
contourpy==1.3.1
crcmod==1.7
cycler==0.12.1
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
ecdsa==0.19.1
ed25519-blake2b==1.4.1
exceptiongroup==1.2.2
fonttools==4.56.0
gunicorn==23.0.0
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
psycopg2==2.9.10
py-sr25519-bindings==0.2.2
pycparser==2.22
pycryptodome==3.22.0
pygame==2.6.1
PyNaCl==1.5.0
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
